+++ ConcernsCaseWork/Service.TRAMS/RecordAcademy/IRecordAcademyService.cs
@@ -1,12 +1,11 @@
 ﻿using System.Collections.Generic;
-using System.Numerics;
 using System.Threading.Tasks;
 
 namespace Service.TRAMS.RecordAcademy
 {
 	public interface IRecordAcademyService
 	{
-		Task<IList<RecordAcademyDto>> GetRecordsAcademyByRecordUrn(BigInteger recordUrn);
+		Task<IList<RecordAcademyDto>> GetRecordsAcademyByRecordUrn(long recordUrn);
 		Task<RecordAcademyDto> PostRecordAcademyByRecordUrn(CreateRecordAcademyDto createRecordAcademyDto);
 		Task<RecordAcademyDto> PatchRecordAcademyByUrn(RecordAcademyDto recordAcademyDto);
 	}
