+++ ConcernsCaseWork/Service.TRAMS/RecordRatingHistory/IRecordRatingHistoryService.cs
@@ -1,13 +1,12 @@
 ﻿using System.Collections.Generic;
-using System.Numerics;
 using System.Threading.Tasks;
 
 namespace Service.TRAMS.RecordRatingHistory
 {
 	public interface IRecordRatingHistoryService
 	{
-		Task<IList<RecordRatingHistoryDto>> GetRecordsRatingHistoryByCaseUrn(BigInteger caseUrn);
-		Task<IList<RecordRatingHistoryDto>> GetRecordsRatingHistoryByRecordUrn(BigInteger recordUrn);
+		Task<IList<RecordRatingHistoryDto>> GetRecordsRatingHistoryByCaseUrn(long caseUrn);
+		Task<IList<RecordRatingHistoryDto>> GetRecordsRatingHistoryByRecordUrn(long recordUrn);
 		Task<RecordRatingHistoryDto> PostRecordRatingHistory(RecordRatingHistoryDto recordRatingHistoryDto);
 	}
 }
