+++ ConcernsCaseWork/Service.TRAMS/RecordSrma/IRecordSrmaService.cs
@@ -1,12 +1,11 @@
 ﻿using System.Collections.Generic;
-using System.Numerics;
 using System.Threading.Tasks;
 
 namespace Service.TRAMS.RecordSrma
 {
 	public interface IRecordSrmaService
 	{
-		Task<IList<RecordSrmaDto>> GetRecordsSrmaByRecordUrn(BigInteger recordUrn);
+		Task<IList<RecordSrmaDto>> GetRecordsSrmaByRecordUrn(long recordUrn);
 		Task<RecordSrmaDto> PostRecordSrmaByRecordUrn(CreateRecordSrmaDto createRecordSrmaDto);
 		Task<RecordSrmaDto> PatchRecordSrmaByUrn(RecordSrmaDto recordSrmaDto);
 	}
