+++ ConcernsCaseWork/Service.TRAMS/RecordWhistleblower/IRecordWhistleblowerService.cs
@@ -1,12 +1,11 @@
 ﻿using System.Collections.Generic;
-using System.Numerics;
 using System.Threading.Tasks;
 
 namespace Service.TRAMS.RecordWhistleblower
 {
 	public interface IRecordWhistleblowerService
 	{
-		Task<IList<RecordWhistleblowerDto>> GetRecordsWhistleblowerByRecordUrn(BigInteger recordUrn);
+		Task<IList<RecordWhistleblowerDto>> GetRecordsWhistleblowerByRecordUrn(long recordUrn);
 		Task<RecordWhistleblowerDto> PostRecordWhistleblowerByRecordUrn(CreateRecordWhistleblowerDto createRecordWhistleblowerDto);
 		Task<RecordWhistleblowerDto> PatchRecordWhistleblowerByUrn(RecordWhistleblowerDto recordWhistleblowerDto);
 	}
