+++ ConcernsCaseWork/Service.TRAMS/Records/IRecordService.cs
@@ -1,12 +1,11 @@
 ﻿using System.Collections.Generic;
-using System.Numerics;
 using System.Threading.Tasks;
 
 namespace Service.TRAMS.Records
 {
 	public interface IRecordService
 	{
-		Task<IList<RecordDto>> GetRecordsByCaseUrn(BigInteger caseUrn);
+		Task<IList<RecordDto>> GetRecordsByCaseUrn(long caseUrn);
 		Task<RecordDto> PostRecordByCaseUrn(CreateRecordDto createRecordDto);
 		Task<RecordDto> PatchRecordByUrn(UpdateRecordDto updateRecordDto);
 	}
